CRUD operations for Gallery

+++ src/main/java/com/example/ArtGallery/repository/GalleryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface GalleryRepository extends JpaRepository<Gallery, Integer> {
+public interface GalleryRepository extends JpaRepository<Gallery, Long> {
     Optional<Gallery> findByName(String datum);
 }
